+++ src/modules/InjectExclusions.ts
@@ -10,6 +10,9 @@ import DefaultData from "../data/IncludeTemplates.json";
 import ExtensionData from "../data/extensions/Includes_Ext.json";
 import * as _ from "lodash";
 import * as fs from "fs";
+import * as filesys from "../utils/FilesystemHelper";
+import * as path from "path";
+import * as vs from "../modules/VSInterface";
 
 export default async function InjectExcludeDefinition(): Promise<void> {
 	return new Promise<void>((resolve, reject) => {
@@ -22,6 +25,7 @@ export default async function InjectExcludeDefinition(): Promise<void> {
 				// console.log(file);
 			}
 		});
+
 		// Get this extension's settings
 		let myconfig = vscode.workspace.getConfiguration("SF");
 		let myretval;
@@ -77,6 +81,44 @@ export default async function InjectExcludeDefinition(): Promise<void> {
 		retval["**.code-workspace"] = true;
 		config.update("exclude", retval, undefined);
 		//#endregion
+
+		// vscode
+		// config = vscode.workspace.getConfiguration(folderpath);
+
+		// vscode.workspace.workspaceFolders?.concat([
+		// 	{
+		// 		uri: vscode.Uri.file(folderpath!),
+		// 		name: "Extensions",
+		// 		index: 1,
+		// 	},
+		// ]);
+		// vscode.workspace.updateWorkspaceFolders(2, 0, {
+		// 	uri: vscode.Uri.file(folderpath!),
+		// 	name: "Extensions",
+		// });
+		// vscode.workspace.updateWorkspaceFolders(2, 0, { uri: new vscode.Uri("") });
+
+		// Adds Extensions tab to workspace
+		let folderpath = filesys.RelativeToAbsolute(
+			"suvam0451.sleeping-forest-ue4",
+			path.join("data", "extensions"),
+		);
+		vscode.workspace.updateWorkspaceFolders(vscode.workspace.workspaceFolders!.length, 0, {
+			uri: vscode.Uri.file(folderpath!),
+			name: "Extensions",
+		});
+
+		// Include asset folders as siderbars in project
+		let choice = vs.GetVSConfig<string[]>("SF", "assetFolders");
+
+		choice.forEach((folder, i) => {
+			if (fs.existsSync(folder)) {
+				vscode.workspace.updateWorkspaceFolders(vscode.workspace.workspaceFolders!.length, 0, {
+					uri: vscode.Uri.file(folder),
+					name: "Stream #" + i,
+				});
+			}
+		});
 
 		resolve();
 	});