+++ src/modules/VSInterface.ts
@@ -55,6 +55,7 @@ export async function PickFolder(): Promise<string> {
 	});
 }
 
+/** Gets vs config */
 export function GetVSConfig<T>(namespace: string, key: string): T {
 	let config = vscode.workspace.getConfiguration(namespace);
 	let retval = config.get<T>(key)!;